+++ rmk-macro/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support more nRF chips: nRF52833, nRF52810, nRF52811
+- Set default GPIO level according to `keybaord.toml`
+
 ## [0.1.5] - 2024-06-08
 
 ### Added